can now verify connection proofs
