+++ frame/ibc/ibc-primitives/Cargo.toml
@@ -22,7 +22,7 @@ base58 = "0.2.0"
 
 [dependencies.ibc]
 git = "https://github.com/ComposableFi/ibc-rs"
-rev = "40ab161e20a3c87c0a50fa152885cf292b8068e4"
+rev = "d8dd882ef4338d6fea2ee4cafb37c25e89f45828"
 default-features = false
 features = ["ics11_beefy"]
